Fix: Use IntTensor not LongTensor for collate_fn compatibility
TypeError: collate_fn expects dtype=int (int32) but got dtype=long (int64)
Changed torch.LongTensor() to torch.IntTensor() in get_text patch
Fixes dataset.py line 238 type mismatch error

+++ utils/amharic_bpe_tokenizer_patch.py
@@ -150,8 +150,9 @@ def apply_global_amharic_bpe_patch():
                 if base_lang in ('am', 'amh'):
                     tokens = self.tokenizer.encode(text, lang)
                     # Convert list to tensor (required by dataset.__getitem__ line 173)
+                    # MUST be IntTensor (int32) not LongTensor - collate_fn expects dtype=int
                     if isinstance(tokens, list):
-                        return torch.LongTensor(tokens)
+                        return torch.IntTensor(tokens)
                     return tokens
                 
                 # For other languages, use original strict check
